requirements.txt
absl-py==2.1.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
Bottleneck==1.3.7
Brotli==1.0.9
cached-property==1.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
comm==0.2.1
contourpy==1.2.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.2.0
executing==0.8.3
fastjsonschema==2.20.0
filelock==3.17.0
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
gast==0.4.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.62.2
h11==0.14.0
h2==4.1.0
h5py==3.12.1
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
imageio==2.33.1
importlib-metadata==7.0.1
importlib-resources==6.4.0
ipykernel==6.28.0
ipython==8.15.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.0
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.2.5
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
keras==3.5.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==18.1.1
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl-fft==1.3.11
mkl-random==1.2.8
mkl-service==2.4.0
ml-dtypes==0.4.1
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook-shim==0.2.4
numexpr==2.8.7
numpy==1.23.5
opencv-python==4.11.0.86
opt-einsum==3.3.0
optree==0.12.1
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
pkgutil-resolve-name==1.3.10
platformdirs==3.10.0
ply==3.11
prettytable==3.5.0
prometheus-client==0.21.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
Pygments==2.15.1
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==75.1.0
sip==6.7.12
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.2.0
tensorboard==2.17.0
tensorboard-data-server==0.7.0
tensorflow==2.17.0
tensorflow-estimator==2.6.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.1.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20240906
typing-extensions==4.11.0
typing-utils==0.1.0
tzdata==2023.3
unicodedata2==15.1.0
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.5
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.44.0
wrapt==1.14.1
zipp==3.17.0
zstandard==0.22.0
